plugins: refactor
