+++ apps/_back02keystone/custom-plugins/tracked.js
@@ -1,7 +1,7 @@
 const { DateTimeUtc } = require('@keystonejs/fields')
 const { AuthedRelationship } = require('@keystonejs/fields-authed-relationship')
 
-const { composeHook } = require('./_common')
+const { composeHook } = require('./utils')
 
 const tracked = ({ createdAtField = 'createdAt', createdByField = 'createdBy', updatedAtField = 'updatedAt', updatedByField = 'updatedBy' } = {}) => ({ fields = {}, hooks = {}, ...rest }) => {
     const datedOptions = {
