+++ apps/_back02keystone/custom-plugins/versioned.js
@@ -1,5 +1,6 @@
 const { Integer } = require('@keystonejs/fields')
-const { composeHook } = require('./_common')
+
+const { composeHook } = require('./utils')
 
 const versioned = ({ versionField = 'v', startBy = 1 } = {}) => ({ fields = {}, hooks = {}, ...rest }) => {
     const versionOptions = {
